Refactor git revision function to static function
